allow object header size to be specified at build time
By default, the object header consists of a single pointer to the
class representing the type of that object and is thus exactly the
size of a native pointer on the target architecture.  This commit
allows additional space to be reserved in the header by changing
ObjectTagInBytes in constants.h.  The additional space can be accessed
with e.g. fieldAtOffset<uintptr_t>(instance, BytesPerWord).